# api/openapi/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This API contains the source for the med.uz app",
        "title": "MED App",
        "contact": {}
    },
    "paths": {
        "/api/appointment": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Creates a new appointment",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Appointment"
                ],
                "summary": "Create Appointment",
                "parameters": [
                    {
                        "description": "Appointment Request",
                        "name": "appointment",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/schemas.AppointmentPayload"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/schemas.ResponseSuccess"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/schemas.ResponseError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/schemas.ResponseError"
                        }
                    }
                }
            }
        },
        "/api/appointment/{id}": {
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Updates an existing appointment",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Appointment"
                ],
                "summary": "Update Appointment",
                "parameters": [
                    {
                        "type": "string",
                        "example": "UUID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Appointment Update Request",
                        "name": "appointment",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/schemas.AppointmentPayload"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/schemas.ResponseSuccess"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/schemas.ResponseError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/schemas.ResponseError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/schemas.ResponseError"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Deletes an existing appointment by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Appointment"
                ],
                "summary": "Delete Appointment",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Appointment ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/schemas.ResponseSuccess"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/schemas.ResponseError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/schemas.ResponseError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/schemas.ResponseError"
                        }
                    }
                }
            }
        },
        "/api/appointment/{id}/status": {
            "patch": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Updates the status of an existing appointment",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Appointment"
                ],
                "summary": "Update Appointment Status",
                "parameters": [
                    {
                        "type": "string",
                        "example": "UUID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "enum": [
                            "pending",
                            "confirmed",
                            "canceled"
                        ],
                        "type": "string",
                        "description": "Appointment Status Update Request",
                        "name": "appointment",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/schemas.ResponseSuccess"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/schemas.ResponseError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/schemas.ResponseError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/schemas.ResponseError"
                        }
                    }
                }
            }
        },
...
        "/api/doctor-appointments/{id}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Retrieves a list of appointments for a specific doctor",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Appointment"
                ],
                "summary": "Get Doctor Appointments",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Doctor ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "search",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/schemas.ManyDoctorAppointment"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/schemas.ResponseError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/schemas.ResponseError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/schemas.ResponseError"
                        }
                    }
                }
            }
        },
...
        "/api/user-appointments/{id}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Retrieves a list of appointments for a specific user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Appointment"
                ],
                "summary": "Get User Appointments",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "search",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/schemas.ManyUserAppointment"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/schemas.ResponseError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/schemas.ResponseError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/schemas.ResponseError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "schemas.AppointmentPayload": {
            "type": "object",
            "required": [
                "appointment_time",
                "doctor_id",
                "user_id"
            ],
            "properties": {
                "appointment_time": {
                    "type": "string",
                    "example": "YYYY-MM-DD H:M:S"
                },
                "doctor_id": {
                    "type": "string",
                    "example": "UUID"
                },
                "user_id": {
                    "type": "string",
                    "example": "UUID"
                }
            }
        },
        "schemas.AppointmentStatus": {
            "type": "string",
            "enum": [
                "pending",
                "confirmed",
                "canceled"
            ],
            "x-enum-varnames": [
                "AppointmentStatusPending",
                "AppointmentStatusConfirmed",
                "AppointmentStatusCanceled"
            ]
        },
...
        "schemas.DoctorAppointment": {
            "type": "object",
            "properties": {
                "appointment_time": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "doctor_info": {
                    "type": "object",
                    "properties": {
                        "experience": {
                            "type": "integer"
                        },
                        "first_name": {
                            "type": "string"
                        },
                        "image": {
                            "type": "string"
                        },
                        "last_name": {
                            "type": "string"
                        },
                        "specialization": {
                            "type": "string"
                        }
                    }
                },
                "id": {
                    "type": "string"
                },
                "s": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/schemas.AppointmentStatus"
                }
            }
        },
...
        "schemas.ManyDoctorAppointment": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "integer"
                },
                "doctor_appointments": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/schemas.DoctorAppointment"
                    }
                }
            }
        },
...
        "schemas.ManyUserAppointment": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "integer"
                },
                "user_appointments": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/schemas.UserAppointment"
                    }
                }
            }
        },
...
        "schemas.NullAppointmentStatus": {
            "type": "object",
            "properties": {
                "appointment_status": {
                    "$ref": "#/definitions/schemas.AppointmentStatus"
                },
                "valid": {
                    "type": "boolean"
                }
            }
        },
        "schemas.ResponseError": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                }
            }
        },
        "schemas.ResponseSuccess": {
            "type": "object",
            "properties": {
                "data": {},
                "status": {
                    "type": "integer"
                }
            }
        },
...
        "schemas.UserAppointment": {
            "type": "object",
            "properties": {
                "appointment_time": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/schemas.NullAppointmentStatus"
                },
                "updated_at": {
                    "type": "string"
                },
                "user_info": {
                    "type": "object",
                    "properties": {
                        "full_name": {
                            "type": "string"
                        },
                        "phone_number": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}